scooper speed set to double

+++ src/main/java/frc/robot/subsystems/Scooper.java
@@ -14,7 +14,7 @@ public class Scooper extends SubsystemBase {
 
   }
 
-  public void setScoop(int speed){
+  public void setScoop(double speed){
     scoopmotor.set(speed);
   }
